Model Refactoring: document model bindings are now working with the help of a projectSyncState kept internally [refs #662, refs #629, refs #661]
The document models have its own properties now instead of just pointing to the project's properties. Therefore they are now observing the project model if their properties should stay in sync with the project. When calling a setter on the document model, the appropriate properties loses it sync mode and is independent of the projects property in the future.

For synced properties 'nil' is stored for the documents property when saving to the project file. This enables the compilable to determine the sync state when loading the project file. At runtime, a projectSyncState-object is held by the DM, keeping informations about the sync state for each synced property. This enables the DM, to unregister all project observers at the projects life cycle end.