disabling TextDidChange notifications while moving lines and firing this notification after finishing line changes [fixes #648]
The problem was that the TextViewController was busy updating the outline view after the first part of line movement was done (e.g. deleting the line or duplicating the line). Hence when firing the change event after the first time, the TVC ignores it and updates the outline view with an old text state.